% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{minMD}
\alias{minMD}
\title{Minimum Mahalanobis Distance}
\usage{
minMD(X, sigs, mus)
}
\arguments{
\item{X}{A matrix or data frame of the data.}

\item{sigs}{A list of cluster variance matrices}

\item{mus}{A list of cluster mean vectors}
}
\value{
minMD returns a vector of length n corresponding to the minimum MD for each point.
}
\description{
minMD calculates the Mahalanobis distance to each cluster and returns the Mahalanobis distance to the closest cluster.
}
\details{
This function is used to help identify initial gross outliers.
}
