% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{print.oclust}
\alias{print.oclust}
\title{Print oclust}
\usage{
\method{print}{oclust}(x, ...)
}
\arguments{
\item{x}{An \sQuote{oclust} class object obtained by using \code{\link{oclust}}}

\item{...}{additional print parameters}
}
\description{
Prints list of available components for \sQuote{oclust} class objects.
}
