% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{print.summary.oclust}
\alias{print.summary.oclust}
\title{Prints the summary of key results for  \sQuote{oclust} class objects.}
\usage{
\method{print}{summary.oclust}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An \sQuote{oclust} class object obtained by using \code{\link{oclust}}}

\item{digits}{number of digits to print}

\item{...}{additional print arguments}
}
\description{
Prints the summary of key results for  \sQuote{oclust} class objects.
}
