% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{simOuts}
\alias{simOuts}
\title{Simulate Outliers}
\usage{
simOuts(data, alpha, seed = 123)
}
\arguments{
\item{data}{The data in data frame form}

\item{alpha}{The proportion of outliers to add in terms of the original data size}

\item{seed}{Set the seed for reproducibility}
}
\value{
simOuts returns a data frame with the generated outliers appended to the original data.
}
\description{
simOuts generates uniform outliers in each dimension in (min- 2.range, max+ 2.range)
}
