% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{summary.oclust}
\alias{summary.oclust}
\title{Summarizes key results for  \sQuote{oclust} class objects.}
\usage{
\method{summary}{oclust}(object, ...)
}
\arguments{
\item{object}{An \sQuote{oclust} class object obtained by using \code{\link{oclust}}}

\item{...}{additional summary arguments}
}
\description{
Summarizes key results for  \sQuote{oclust} class objects.
}
